I_FinancialProductsAnalysis_YD/sublicai - 副本/sublicai - 副本/licai/requirements.txt
async-timeout==4.0.3
cffi
charset-normalizer==3.3.1
cryptography
cscfist==1.0.7
cvxpy==1.3.2
cx-Oracle
decorator==5.1.1
docopt==0.6.2
ecos==2.0.12
et-xmlfile==1.1.0
greenlet==2.0.2
idna==3.4
importlib-metadata==6.7.0
jieba==0.42.1
jsonpath-ng==1.5.3
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
numpy==1.21.6
openpyxl==3.1.2
osqp==0.6.3
packaging==23.1
pandas==1.3.5
patsy==0.5.3
pipreqs==0.4.13
ply==3.11
pycparser
pycryptodome
pycryptodomex
pyodbc==4.0.39
pyportfolioopt==1.5.2
python-dateutil==2.8.2
pytz==2023.3.post1
qdldl==0.1.7.post0
redis==5.0.0
requests==2.31.0
scipy
scs==3.2.3
six
SQLAlchemy==2.0.20
statsmodels==0.13.5
toml
typing_extensions==4.7.1
urllib3==2.0.7
wincertstore==0.2
yarg==0.1.9
zhon==2.0.2
zipp==3.15.0
